Rio
Rio it seems that the part without bullets was just duplicated for some reason and thats why it looks the way it does I have tried to erase the duplicated portion and it won let me. After your first suggestion that I change it I did so but it alsoduplicated the entire section for some reason not sure why.